refactor: 웨이팅 관련 API 엔드포인트 이동 및 Service 인터페이스 정리
* **refactor: 장소 관련 웨이팅 API를 `WaitingService`에서 `PlaceService`로 이동**
    * `places/{placeId}/waiting-status` (조회) 및 `places/{placeId}/waitings` (등록) API를 도메인 성격에 맞춰 `PlaceService`로 재배치했습니다.
    * `WaitingService`에는 본인의 웨이팅 정보 조회 및 취소 API만 남도록 정리했습니다.

* **refactor: `WaitingRemoteDataSourceImpl` 내 서비스 의존성 수정**
    * `PlaceService`를 주입받아 이동된 웨이팅 관련 API(`fetchPlaceWaiting`, `registerWaiting`)를 호출하도록 로직을 수정했습니다.
    * 불필요해진 임포트 및 관련 요청/응답 모델 참조를 정리했습니다.
