+++ composeApp/src/commonMain/kotlin/com/daedan/festabook/data/datasource/remote/waiting/WaitingRemoteDataSourceImpl.kt
@@ -4,6 +4,7 @@ import com.daedan.festabook.data.datasource.remote.ApiResult
 import com.daedan.festabook.data.model.request.WaitingRegisterRequest
 import com.daedan.festabook.data.model.response.waiting.MyWaitingResponse
 import com.daedan.festabook.data.model.response.waiting.PlaceWaitingResponse
+import com.daedan.festabook.data.service.PlaceService
 import com.daedan.festabook.data.service.WaitingService
 import dev.zacsweers.metro.AppScope
 import dev.zacsweers.metro.ContributesBinding
@@ -13,16 +14,17 @@ import dev.zacsweers.metro.Inject
 @Inject
 class WaitingRemoteDataSourceImpl(
     private val waitingService: WaitingService,
+    private val placeService: PlaceService,
 ) : WaitingRemoteDataSource {
     override suspend fun fetchMyWaiting(): ApiResult<MyWaitingResponse> = ApiResult.toApiResult { waitingService.fetchMyWaiting() }
 
     override suspend fun fetchPlaceWaiting(placeId: Long): ApiResult<PlaceWaitingResponse> =
-        ApiResult.toApiResult { waitingService.fetchPlaceWaiting(placeId) }
+        ApiResult.toApiResult { placeService.fetchPlaceWaiting(placeId) }
 
     override suspend fun cancelWaiting(waitingId: Long): ApiResult<Unit> = ApiResult.toApiResult { waitingService.cancelWaiting(waitingId) }
 
     override suspend fun registerWaiting(
         placeId: Long,
         request: WaitingRegisterRequest,
-    ): ApiResult<MyWaitingResponse> = ApiResult.toApiResult { waitingService.registerWaiting(placeId, request) }
+    ): ApiResult<MyWaitingResponse> = ApiResult.toApiResult { placeService.registerWaiting(placeId, request) }
 }
