+++ composeApp/src/commonMain/kotlin/com/daedan/festabook/data/service/PlaceService.kt
@@ -1,11 +1,16 @@
 package com.daedan.festabook.data.service
 
+import com.daedan.festabook.data.model.request.WaitingRegisterRequest
 import com.daedan.festabook.data.model.response.place.PlaceDetailResponse
 import com.daedan.festabook.data.model.response.place.PlaceGeographyResponse
 import com.daedan.festabook.data.model.response.place.PlaceResponse
 import com.daedan.festabook.data.model.response.place.TimeTagResponse
+import com.daedan.festabook.data.model.response.waiting.MyWaitingResponse
+import com.daedan.festabook.data.model.response.waiting.PlaceWaitingResponse
 import de.jensklingenberg.ktorfit.Response
+import de.jensklingenberg.ktorfit.http.Body
 import de.jensklingenberg.ktorfit.http.GET
+import de.jensklingenberg.ktorfit.http.POST
 import de.jensklingenberg.ktorfit.http.Path
 
 interface PlaceService {
@@ -22,4 +27,15 @@ interface PlaceService {
 
     @GET("places/geographies")
     suspend fun fetchPlaceGeographies(): Response<List<PlaceGeographyResponse>>
+
+    @GET("places/{placeId}/waiting-status")
+    suspend fun fetchPlaceWaiting(
+        @Path("placeId") placeId: Long,
+    ): Response<PlaceWaitingResponse>
+
+    @POST("places/{placeId}/waitings")
+    suspend fun registerWaiting(
+        @Path("placeId") placeId: Long,
+        @Body request: WaitingRegisterRequest,
+    ): Response<MyWaitingResponse>
 }
